+++ changelog.md
@@ -12,6 +12,7 @@
 - difference renamed into negation
 - added table.union, table.deepflat, table.equal
 - some operator overloads
+- added lshift, rshift and shift
 
 ## (0.6) 01/06/2018:
 - new table operator: add